webcam/requirements.txt
Pillow==10.0.0
reflex==0.3.4a2

